- adapt to alloy from ethers-rs - bump version to 0.4.0-alpha.1

+++ crates/arpa-node/src/context/chain/mod.rs
@@ -25,7 +25,7 @@ pub trait Chain<
     type RandomnessResultCaches;
     type ChainIdentity;
 
-    fn id(&self) -> usize;
+    fn id(&self) -> u64;
 
     fn description(&self) -> &str;
