+++ crates/arpa-node/src/event/provider_reconnection.rs
@@ -3,11 +3,11 @@ use crate::subscriber::DebuggableEvent;
 
 #[derive(Clone, Debug)]
 pub struct ProviderReconnection {
-    pub chain_id: usize,
+    pub chain_id: u64,
 }
 
 impl ProviderReconnection {
-    pub fn new(chain_id: usize) -> Self {
+    pub fn new(chain_id: u64) -> Self {
         ProviderReconnection { chain_id }
     }
 }
